CIV-7117 Recfor Listing for hearing Respondent Noft. (#905)
* CIV-7117 Rector Listing for hearing Respondent Noft.

* CIV-7117 Notification hearing scheduled

* CIV-7117 Refactor helper class

* CIV-7117 Remove unused method

---------
