+++ src/main/java/uk/gov/hmcts/reform/civil/service/JudicialNotificationService.java
@@ -23,6 +23,7 @@ import static uk.gov.hmcts.reform.civil.helpers.DateFormatHelper.DATE;
 import static uk.gov.hmcts.reform.civil.helpers.DateFormatHelper.JUDICIAL_FORMATTER;
 import static uk.gov.hmcts.reform.civil.utils.JudicialDecisionNotificationUtil.areRespondentSolicitorsPresent;
 import static uk.gov.hmcts.reform.civil.utils.JudicialDecisionNotificationUtil.isApplicationCloaked;
+import static uk.gov.hmcts.reform.civil.utils.JudicialDecisionNotificationUtil.isWithNotice;
 import static uk.gov.hmcts.reform.civil.utils.JudicialDecisionNotificationUtil.notificationCriterion;
 import static uk.gov.hmcts.reform.civil.utils.JudicialDecisionNotificationUtil.requiredGAType;
 
@@ -323,7 +324,10 @@ public class JudicialNotificationService implements NotificationData {
     private void applicationListForHearing(CaseData caseData, String solicitorType) {
 
         if (solicitorType.equals(RESPONDENT)) {
-            if (areRespondentSolicitorsPresent(caseData)) {
+            /*
+            * Respondent should receive notification only if it's with notice application
+            *  */
+            if (isWithNotice(caseData) && areRespondentSolicitorsPresent(caseData)) {
                 sendEmailToRespondent(
                     caseData,
                     notificationProperties.getJudgeListsForHearingRespondentEmailTemplate()
